refactor: renderer architecture — typed events, stream extraction, CSS split
- Replace untyped CustomEvent bus (24 sites) with typed OrchidEventMap registry
  (utils/events.ts) providing compile-time safety for event names and payloads
- Extract 10 pure data-transform functions from ChatStream.tsx (1006→420 lines)
  into utils/stream-building.ts
- Split components.css (2249 lines) by surface area into components-chat.css,
  components-config.css, components-session.css
- Update contract tests to match relocated source patterns
