+++ electron/src/renderer/components/Onboarding/OnboardingScreen.tsx
@@ -14,6 +14,7 @@ import {
 } from '../../../shared/mcp/recommended-servers';
 import { useProviders } from '../../hooks/useProviders';
 import { useFocusTrap } from '../../keyboard';
+import { emitOrchidEvent } from '../../utils/events';
 import { isEmbeddingModel, isTextGenerationModel } from '../../utils/models';
 import {
   providerModelOptionDisplayName,
@@ -204,9 +205,7 @@ export function OnboardingScreen({ isOpen, onComplete, onSkip }: OnboardingScree
   }, [isOpen, providers.modelOptions, connectionSignature]);
 
   const restoreConfigTheme = useCallback(() => {
-    window.dispatchEvent(new CustomEvent('orchid:set-theme', {
-      detail: { theme: configTheme, persist: false },
-    }));
+    emitOrchidEvent('orchid:set-theme', { theme: configTheme, persist: false });
   }, [configTheme]);
 
   const markCompleteAndClose = useCallback(async (mode: 'skip' | 'finish') => {
@@ -226,9 +225,7 @@ export function OnboardingScreen({ isOpen, onComplete, onSkip }: OnboardingScree
       if (mode === 'skip') {
         restoreConfigTheme();
         await window.orchid.config.save({ updates: { has_completed_onboarding: true } });
-        window.dispatchEvent(new CustomEvent('orchid:config-updated', {
-          detail: { has_completed_onboarding: true },
-        }));
+        emitOrchidEvent('orchid:config-updated', { has_completed_onboarding: true });
         onSkip();
         return;
       }
@@ -250,15 +247,10 @@ export function OnboardingScreen({ isOpen, onComplete, onSkip }: OnboardingScree
       }
 
       await window.orchid.config.save({ updates });
-      window.dispatchEvent(new CustomEvent('orchid:config-updated', { detail: updates }));
-      window.dispatchEvent(new CustomEvent('orchid:set-theme', {
-        detail: { theme, persist: false },
-      }));
+      emitOrchidEvent('orchid:config-updated', updates);
+      emitOrchidEvent('orchid:set-theme', { theme, persist: false });
       if (defaultModel) {
-        window.dispatchEvent(new CustomEvent<{ selection: ModelSelection }>(
-          'orchid:provider-selection-created',
-          { detail: { selection: defaultModel } },
-        ));
+        emitOrchidEvent('orchid:provider-selection-created', { selection: defaultModel });
       }
       onComplete();
     } catch (error) {
@@ -299,9 +291,7 @@ export function OnboardingScreen({ isOpen, onComplete, onSkip }: OnboardingScree
 
   const handleThemeChange = useCallback((name: ThemeName) => {
     setTheme(name);
-    window.dispatchEvent(new CustomEvent('orchid:set-theme', {
-      detail: { theme: name, persist: false },
-    }));
+    emitOrchidEvent('orchid:set-theme', { theme: name, persist: false });
   }, []);
 
   const handlePickProject = useCallback(async () => {
