+++ electron/src/renderer/components/Preferences/SkillsTab.tsx
@@ -10,6 +10,7 @@ import type {
   DefinitionsListResult,
   ManagedSkill,
 } from '../../../shared/types/definitions';
+import { onOrchidEvent } from '../../utils/events';
 import { Alert } from '../ui/Alert';
 import { Button } from '../ui/Button';
 import { ConfigCard } from '../ui/ConfigCard';
@@ -67,13 +68,11 @@ export function SkillsTab({ data, onReload, lockedScope }: SkillsTabProps) {
   const [saving, setSaving] = useState(false);
 
   useEffect(() => {
-    const clear = () => {
+    return onOrchidEvent('orchid:definitions-workspace-changed', () => {
       setEditingKey(null);
       setIsAdding(false);
       setForm(null);
-    };
-    window.addEventListener('orchid:definitions-workspace-changed', clear);
-    return () => window.removeEventListener('orchid:definitions-workspace-changed', clear);
+    });
   }, []);
 
   const skills = data.skills;
