+++ electron/src/renderer/styles/components-session.css
@@ -1,0 +1,127 @@
+@layer orchid {
+  /* Context bar geometry (dynamic segment colors stay inline exceptions). */
+  .context-stacked-bar {
+    @apply flex w-full overflow-hidden rounded-sm bg-base-300;
+    height: 10px;
+    margin: 4px 0 8px;
+    box-shadow: inset 0 0 0 1px color-mix(in srgb, var(--color-base-content) 10%, transparent);
+  }
+
+  .context-stacked-bar-segment {
+    @apply block h-full min-w-0;
+    transition: width 180ms ease;
+  }
+
+  .context-stacked-bar-segment + .context-stacked-bar-segment {
+    border-left: 1px solid color-mix(in srgb, var(--color-base-100) 45%, transparent);
+  }
+
+  /* Compact mock-style collapse (inspector panels).
+     Surface steps toward the base surface — recessed in dark themes,
+     lifted toward white in light themes. Overlay color is reserved for
+     shadows and veils, never used as a surface tint. */
+  .mock-collapse {
+    margin-bottom: 4px;
+    border: 1px solid var(--border-default);
+    border-radius: 5px;
+    background: color-mix(in srgb, var(--bg-secondary) 84%, var(--bg-primary));
+    overflow: hidden;
+  }
+  .mock-collapse .mock-collapse-title {
+    min-height: 30px;
+    padding: 6px 8px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 6px;
+    font-weight: 600;
+    font-size: 12px;
+    cursor: pointer;
+    text-align: left;
+    width: 100%;
+    border: none;
+    background: transparent;
+    color: inherit;
+    font-family: inherit;
+    transition: background-color var(--transition-fast, 150ms ease);
+  }
+  .mock-collapse .mock-collapse-title:hover {
+    filter: brightness(1.06);
+  }
+  .mock-collapse .mock-collapse-content {
+    padding: 6px 8px;
+    font-size: 12px;
+  }
+
+  /* ── Panel-variant context legend rows ─────────────────────────────────── */
+
+  .context-panel-list {
+    display: flex;
+    flex-direction: column;
+    gap: 6px;
+    margin: 0;
+    padding: 0;
+  }
+  .context-panel-row {
+    @apply flex items-center justify-between gap-3;
+    padding: 8px 10px;
+    border-radius: 7px;
+    background: color-mix(in srgb, var(--text-primary) 2%, transparent);
+    border: 1px solid color-mix(in srgb, var(--text-primary) 4%, transparent);
+  }
+  .context-panel-row:hover {
+    background: color-mix(in srgb, var(--accent-primary) 8%, transparent);
+    border-color: color-mix(in srgb, var(--accent-primary) 16%, transparent);
+  }
+  .context-panel-row-left {
+    display: inline-flex;
+    align-items: center;
+    gap: 8px;
+    min-width: 0;
+  }
+  .context-panel-swatch {
+    width: 8px;
+    height: 8px;
+    border-radius: 2px;
+    flex-shrink: 0;
+    box-shadow: 0 0 0 1px color-mix(in srgb, var(--overlay-color) 25%, transparent);
+  }
+  .context-panel-label {
+    font-size: 12px;
+    font-weight: 600;
+    color: var(--text-primary);
+    line-height: 1.2;
+  }
+  .context-panel-row-right {
+    display: inline-flex;
+    align-items: baseline;
+    gap: 8px;
+    flex-shrink: 0;
+    font-size: 11px;
+    line-height: 1.2;
+  }
+  .context-panel-tokens {
+    color: var(--text-primary);
+    font-weight: 500;
+    font-variant-numeric: tabular-nums;
+  }
+  .context-panel-pct {
+    color: var(--text-secondary);
+    min-width: 2.4em;
+    text-align: right;
+    font-variant-numeric: tabular-nums;
+  }
+  /* ── Compact badge overrides ───────────────────────────────────────────── */
+
+  /* Compact badges in dense chrome (still keep horizontal padding). */
+  .session-item .badge,
+  .session-project-toggle .badge,
+  .inspector-row .badge,
+  .mock-collapse .badge,
+  .orchid-tool-block .badge {
+    min-height: 16px;
+    padding: 2px 7px;
+    font-size: 9px;
+    line-height: 1.2;
+  }
+}
