+++ electron/src/renderer/styles/index.css
@@ -4,7 +4,7 @@
  * Import order matters for cascade processing:
  * 1. Tailwind (theme/base/components/utilities layers)
  * 2. Orchid primitive engine (component classes — joins `@layer components`)
- * 3. Shared orchid-* composites
+ * 3. Shared orchid-* composites (base, then chat/session/config surface splits)
  * 4. Shared motion vocabulary
  * 5. Markdown / highlight exceptions
  * 6. Browser/runtime exceptions
@@ -18,6 +18,9 @@
 @import "tailwindcss";
 @import "./primitives.css";
 @import "./components.css";
+@import "./components-chat.css";
+@import "./components-session.css";
+@import "./components-config.css";
 @import "./shell.css";
 @import "./motion.css";
 @import "./markdown.css";
